Rydder litt i chatbot-kode
